Refactor the codebase structure
- Following Django best-practices
- Simplify the layout
- Use proper naming convention
