+++ requirements.txt
@@ -1,7 +1,10 @@
+setuptools==36.5.0
+
 beautifulsoup4==4.6.0
-lxml==3.8.0
-django==1.11.4
-djangorestframework==3.6
+lxml==4.0.0
+django==1.11.5
+djangorestframework==3.6.4
 
 # Tests
-pytest==3.1.3
+pytest==3.2.2
+pycodestyle==2.3.1
