+++ vulnerabilities/data_dump.py
@@ -21,10 +21,10 @@
 #  VulnerableCode is a free software code scanning tool from nexB Inc. and others.
 #  Visit https://github.com/nexB/vulnerablecode/ for support and download.
 
-from vulncode_app.models import ImpactedPackage
-from vulncode_app.models import Package
-from vulncode_app.models import Vulnerability
-from vulncode_app.models import VulnerabilityReference
+from vulnerabilities.models import ImpactedPackage
+from vulnerabilities.models import Package
+from vulnerabilities.models import Vulnerability
+from vulnerabilities.models import VulnerabilityReference
 
 
 def debian_dump(extract_data):
