+++ vulnerabilities/tests/test_api_data.py
@@ -23,7 +23,7 @@
 
 import json
 
-from vulncode_app.api_data import extract_fields
+from vulnerabilities.api_data import extract_fields
 
 
 test_data = """
